+++ test/unit/events.js
@@ -14,6 +14,32 @@ test('should add and remove an event listener to an element', function(){
   vjs.trigger(el, 'click'); // No click should happen.
 });
 
+test('should add and remove multiple event listeners to an element with a single call', function(){
+  expect(6);
+
+  var el = document.createElement('div');
+  var listener = function(){
+    ok(true, 'Callback triggered');
+  };
+
+  vjs.on(el, ['click', 'event1', 'event2'], listener);
+
+  vjs.trigger(el, 'click');
+  vjs.trigger(el, 'click');
+  vjs.off(el, 'click', listener);
+  vjs.trigger(el, 'click'); // No click should happen.
+
+  vjs.trigger(el, 'event1');
+  vjs.trigger(el, 'event1');
+  vjs.off(el, 'event1', listener);
+  vjs.trigger(el, 'event1'); // No event1 should happen.
+
+  vjs.trigger(el, 'event2');
+  vjs.trigger(el, 'event2');
+  vjs.off(el, 'event2', listener);
+  vjs.trigger(el, 'event2'); // No event2 should happen.
+});
+
 test('should remove all listeners of a type', function(){
   var el = document.createElement('div');
   var clicks = 0;
@@ -34,6 +60,30 @@ test('should remove all listeners of a type', function(){
   vjs.trigger(el, 'click'); // No click should happen.
 
   ok(clicks === 2, 'no click listeners fired');
+});
+
+test('should remove all listeners of an array of types', function(){
+  var el = document.createElement('div');
+  var calls = 0;
+  var listener = function(){
+    calls++;
+  };
+  var listener2 = function(){
+    calls++;
+  };
+
+  vjs.on(el, ['click', 'event1'], listener);
+  vjs.on(el, ['click', 'event1'], listener2);
+  vjs.trigger(el, 'click'); // 2 calls
+  vjs.trigger(el, 'event1'); // 2 calls
+
+  ok(calls === 4, 'both click listeners fired');
+
+  vjs.off(el, ['click', 'event1']);
+  vjs.trigger(el, 'click'); // No click should happen.
+  vjs.trigger(el, 'event1'); // No event1 should happen.
+
+  ok(calls === 4, 'no event listeners fired');
 });
 
 test('should remove all listeners from an element', function(){
@@ -71,6 +121,23 @@ test('should listen only once', function(){
   vjs.one(el, 'click', listener);
   vjs.trigger(el, 'click'); // 1 click
   vjs.trigger(el, 'click'); // No click should happen.
+});
+
+test( 'should listen only once in multiple events from a single call', function(){
+  expect(3);
+
+  var el = document.createElement('div');
+  var listener = function(){
+    ok(true, 'Callback Triggered');
+  };
+
+  vjs.one(el, ['click', 'event1', 'event2'], listener);
+  vjs.trigger(el, 'click'); // 1 click
+  vjs.trigger(el, 'click'); // No click should happen.
+  vjs.trigger(el, 'event1'); // event1 must be handled.
+  vjs.trigger(el, 'event1'); // No event1 should be handled.
+  vjs.trigger(el, 'event2'); // event2 must be handled. 
+  vjs.trigger(el, 'event2'); // No event2 should be handled.
 });
 
 test('should stop immediate propagtion', function(){