+++ CHANGELOG.md
@@ -12,6 +12,7 @@ CHANGELOG
 * Exported missing Player API methods (remainingTime, supportsFullScreen, enterFullWindow, exitFullWindow, preload) ([view](https://github.com/videojs/video.js/pull/1328))
 * Added a base for running saucelabs tests from grunt ([view](https://github.com/videojs/video.js/pull/1215))
 * Added additional browsers for saucelabs testing ([view](https://github.com/videojs/video.js/pull/1216))
+* Added support for listening to multiple events through a types array ([view](https://github.com/videojs/video.js/pull/1231))
 
 --------------------